requirement.txt
streamlit==1.32.0
tensorflow==2.15.0
pillow==10.2.0
numpy==1.24.3
matplotlib==3.7.1

